Move middleware usage to the proper place
