+++ config/environments/production.rb
@@ -154,6 +154,8 @@ Rails.application.configure do
                           ENV["HEROKU_APP_URL"] => ENV["APP_DOMAIN"]
   end
 
+  config.middleware.use Middlewares::SetCookieDomain
+
   # Inserts middleware to perform automatic connection switching.
   # The `database_selector` hash is used to pass options to the DatabaseSelector
   # middleware. The `delay` is used to determine how long to wait after a write
